policy_holder_bot/requirements.txt
ic-py==1.0.1
psycopg2==2.9.10
python-telegram-bot==20.3
python-dotenv==1.0.1
sqlalchemy==2.0.37
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==4.1.0